requirements.txt
ag-ui-protocol==0.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
argcomplete==3.6.2
attrs==25.3.0
boto3==1.40.42
botocore==1.40.42
cachetools==6.2.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cohere==5.18.0
colorama==0.4.6
cryptography==46.0.1
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
executing==2.2.1
fastapi==0.118.0
fastavro==1.12.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
genai-prices==0.0.29
google-auth==2.41.1
google-genai==1.39.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.14.0
groq==0.26.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
invoke==2.2.0
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.25
langchain-core==0.3.65
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
logfire==4.10.0
logfire-api==4.10.0
lxml==6.0.2
markdown-it-py==4.0.0
mcp==1.15.0
mdurl==0.1.2
mistralai==1.9.10
multidict==6.6.4
nexus-rpc==1.1.0
openai==2.0.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.10.18
packaging==24.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-ai==1.0.12
pydantic-ai-slim==1.0.12
pydantic-evals==1.0.12
pydantic-graph==1.0.12
pydantic-settings==2.6.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.0
pyperclip==1.11.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==3.0.2
starlette==0.48.0
temporalio==1.18.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
types-protobuf==6.32.1.20250918
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
uvicorn==0.37.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
